final-pjt/final-pjt-back/requirements.txt
asgiref==3.3.4
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.13
django-allauth==0.52.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==3.3
oauthlib==3.2.2
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.1
sqlparse==0.4.1
urllib3==1.26.7



asgiref==3.3.4
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.13
django-allauth==0.52.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==3.3
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.1
sqlparse==0.4.1
urllib3==1.26.7


asgiref==3.3.4
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.13
django-allauth==0.52.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==3.3
joblib==1.2.0
numpy==1.24.3
oauthlib==3.2.2
pandas==2.0.1
Pillow==9.5.0
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.1
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==1.26.7
